+++ example/lib/feedback_functions.dart
@@ -1,11 +1,5 @@
-import 'package:feedback/src/user_feedback.dart';
+import 'package:feedback/feedback.dart';
 import 'package:flutter/material.dart';
-
-/// Function which gets called when the user submits his feedback.
-/// [feedback] is the user generated feedback. A string, by default.
-/// [screenshot] is a raw png encoded image.
-/// [OnFeedbackCallback] should cast [feedback] to the appropriate type.
-typedef OnFeedbackCallback = void Function(UserFeedback);
 
 /// Prints the given feedback to the console.
 /// This is useful for debugging purposes.
@@ -15,7 +9,7 @@ void consoleFeedbackFunction(
 ) {
   print('Feedback text:');
   print(feedback.text);
-  print('Size of image: ${feedback.screenshot?.length}');
+  print('Size of image: ${feedback.screenshot.length}');
   if (feedback.extra != null) {
     print('Extras: ${feedback.extra!.toString()}');
   }
@@ -37,13 +31,12 @@ void alertFeedbackFunction(
             mainAxisSize: MainAxisSize.min,
             children: [
               if (feedback.extra != null) Text(feedback.extra!.toString()),
-              if (feedback.screenshot != null)
-                Image.memory(
-                  feedback.screenshot!,
-                  height: 600,
-                  width: 500,
-                  fit: BoxFit.contain,
-                ),
+              Image.memory(
+                feedback.screenshot,
+                height: 600,
+                width: 500,
+                fit: BoxFit.contain,
+              ),
             ],
           ),
         ),