+++ lib/src/user_feedback.dart
@@ -1,4 +1,5 @@
 import 'dart:typed_data';
+import 'package:feedback/src/better_feedback.dart';
 
 /// Container for the feedback of the user.
 class UserFeedback {
@@ -8,13 +9,14 @@ class UserFeedback {
     this.extra,
   });
 
-  /// Written feedback of the user
+  /// The user's written feedback
   final String text;
 
   /// A raw png encoded screenshot of the app. Hopefully annotated by helpful
-  /// drawings of the user.
-  final Uint8List? screenshot;
+  /// drawings by the user.
+  final Uint8List screenshot;
 
-  /// Additional information. For user by custom Feedback Builders.
+  /// Additional information.
+  /// For use by custom [BetterFeedback.feedbackBuilder].
   final Map<String, dynamic>? extra;
 }