+++ test/test_app.dart
@@ -75,7 +75,7 @@ class MyTestPageState extends State<MyTestPage> {
               key: const Key('open_feedback'),
               child: const Text('open feedback'),
               onPressed: () {
-                BetterFeedback.of(context)!.show(widget.onFeedback ?? (_) {});
+                BetterFeedback.of(context).show(widget.onFeedback ?? (_) {});
               },
             ),
             TextButton(